[IMP] web_advanced_search: Move legacy widget overrides to specific folder
