+++ web_advanced_search/static/src/js/legacy/CustomFilterItem.esm.js
@@ -2,9 +2,10 @@
 
 import {patch} from "@web/core/utils/patch";
 import CustomFilterItem from "web.CustomFilterItem";
-import {RecordPicker} from "./RecordPicker.esm";
+import {RecordPicker} from "../RecordPicker.esm";
 
-patch(CustomFilterItem.prototype, "web_advanced_search.CustomFilterItem", {
+
+patch(CustomFilterItem.prototype, "web_advanced_search.legacy.CustomFilterItem", {
     /**
      * Ideally we'd want this in setup, but CustomFilterItem does its initialization
      * in the constructor, which can't be patched.
@@ -75,7 +76,7 @@ patch(CustomFilterItem.prototype, "web_advanced_search.CustomFilterItem", {
     },
 });
 
-patch(CustomFilterItem, "web_advanced_search.CustomFilterItem", {
+patch(CustomFilterItem, "web_advanced_search.legacy.CustomFilterItem", {
     components: {
         ...CustomFilterItem.components,
         RecordPicker,
