+++ web_advanced_search/static/src/js/legacy/FilterMenu.esm.js
@@ -2,9 +2,10 @@
 
 import {patch} from "@web/core/utils/patch";
 import FilterMenu from "web.FilterMenu";
-import AdvancedFilterItem from "./AdvancedFilterItem.esm";
+import AdvancedFilterItem from "../AdvancedFilterItem.esm";
 
-patch(FilterMenu, "web_advanced_search.FilterMenu", {
+
+patch(FilterMenu, "web_advanced_search.legacy.FilterMenu", {
     components: {
         ...FilterMenu.components,
         AdvancedFilterItem,
